+++ src/main/scala/com/persona/service/account/AccountDAO.scala
@@ -1,0 +1,5 @@
+package com.persona.service.account
+
+class AccountDAO {
+
+}
