+++ src/main/scala/com/persona/service/account/AccountService.scala
@@ -1,0 +1,5 @@
+package com.persona.service.account
+
+class AccountService {
+
+}
