+++ src/main/scala/com/persona/service/authentication/PersonaAuthService.scala
@@ -1,0 +1,7 @@
+package com.persona.service.authentication
+
+class PersonaAuthService {
+  def authenticate(basicAuth: BasicAuth) = {
+    s"Hello, ${basicAuth.id}!"
+  }
+}
