+++ src/main/scala/com/persona/service/authentication/facebook/FacebookAuthService.scala
@@ -1,0 +1,7 @@
+package com.persona.service.authentication.facebook
+
+class FacebookAuthService {
+
+  def authenticate = "Logging in to Facebook"
+
+}
