+++ src/main/scala/com/persona/service/authentication/google/GoogleAuthService.scala
@@ -1,0 +1,7 @@
+package com.persona.service.authentication.google
+
+class GoogleAuthService {
+
+  def authenticate = "Logging in to Google"
+
+}
