+++ src/main/scala/com/persona/service/authorization/AuthorizationService.scala
@@ -1,0 +1,9 @@
+package com.persona.service.authorization
+
+import com.nimbusds.jose.jwk.JWKSet
+
+class AuthorizationService {
+
+  def jwks: JWKSet = new JWKSet
+
+}
