+++ src/main/scala/com/persona/service/authorization/PersonaIdToken.scala
@@ -1,0 +1,5 @@
+package com.persona.service.authorization
+
+class PersonaIdToken {
+
+}
