+++ src/main/scala/com/persona/service/authorization/PersonaIdTokenGenerator.scala
@@ -1,0 +1,7 @@
+package com.persona.service.authorization
+
+class PersonaIdTokenGenerator {
+
+
+
+}
