+++ src/main/scala/com/persona/service/authorization/PersonaIdTokenVerifier.scala
@@ -1,0 +1,7 @@
+package com.persona.service.authorization
+
+class PersonaIdTokenVerifier {
+
+  def verify(token: PersonaIdToken): Boolean = true
+
+}
