+++ src/main/scala/com/persona/util/actor/ActorWrapper.scala
@@ -1,0 +1,9 @@
+package com.persona.util.actor
+
+import akka.actor.{ActorRef, PoisonPill}
+
+class ActorWrapper(protected val actor: ActorRef) {
+
+  def stop(): Unit = actor ! PoisonPill
+
+}
