+++ src/main/scala/com/persona/util/openid/OpenIdDiscoveryDocument.scala
@@ -1,0 +1,3 @@
+package com.persona.util.openid
+
+case class OpenIdDiscoveryDocument(jwksUri: String)
