--- src/main/scala/persona/service/account/AccountDAO.scala
@@ -1,5 +1,0 @@
-package persona.service.account
-
-class AccountDAO {
-
-}
