--- src/main/scala/persona/service/account/AccountService.scala
@@ -1,5 +1,0 @@
-package persona.service.account
-
-class AccountService {
-
-}
