--- src/main/scala/persona/service/authentication/PersonaAuthService.scala
@@ -1,7 +1,0 @@
-package persona.service.authentication
-
-class PersonaAuthService {
-  def authenticate(basicAuth: BasicAuth) = {
-    s"Hello, ${basicAuth.id}!"
-  }
-}
