--- src/main/scala/persona/service/authentication/facebook/FacebookAuthService.scala
@@ -1,7 +1,0 @@
-package persona.service.authentication.facebook
-
-class FacebookAuthService {
-
-  def authenticate = "Logging in to Facebook"
-
-}
