--- src/main/scala/persona/service/authentication/google/GoogleAuthService.scala
@@ -1,7 +1,0 @@
-package persona.service.authentication.google
-
-class GoogleAuthService {
-
-  def authenticate = "Logging in to Google"
-
-}
