--- src/main/scala/persona/service/token/ActorWrapper.scala
@@ -1,9 +1,0 @@
-package persona.service.token
-
-import akka.actor.{ActorRef, PoisonPill}
-
-class ActorWrapper(protected val actor: ActorRef) {
-
-  def stop(): Unit = actor ! PoisonPill
-
-}
