--- src/main/scala/persona/service/token/OpenIdDiscoveryDocument.scala
@@ -1,3 +1,0 @@
-package persona.service.token
-
-case class OpenIdDiscoveryDocument(jwksUri: String)
