--- src/main/scala/persona/service/token/PersonaIdToken.scala
@@ -1,5 +1,0 @@
-package persona.service.token
-
-class PersonaIdToken {
-
-}
