--- src/main/scala/persona/service/token/PersonaIdTokenGenerator.scala
@@ -1,7 +1,0 @@
-package persona.service.token
-
-class PersonaIdTokenGenerator {
-
-
-
-}
