--- src/main/scala/persona/service/token/PersonaIdTokenVerifier.scala
@@ -1,7 +1,0 @@
-package persona.service.token
-
-class PersonaIdTokenVerifier {
-
-  def verify(token: PersonaIdToken): Boolean = true
-
-}
