--- src/main/scala/persona/service/token/TokenService.scala
@@ -1,9 +1,0 @@
-package persona.service.token
-
-import com.nimbusds.jose.jwk.JWKSet
-
-class TokenService {
-
-  def jwks: JWKSet = new JWKSet
-
-}
